+++ src/entryIDs.ts
@@ -5,3 +5,4 @@ export const clientRegisterPrivacyPolicyPageID = '6a7fVb49Zf79FTetXflVFL';
 export const indexPageID = 'rXFgpak6HKjCuUXjFo9KW';
 export const managementPageID = '3rPdaUCw3nKo73b5Z6thWW';
 export const marketingRegisterPrivacyPolicyPageID = '2B8WVOvBXdHmLHeBFx381E';
+export const portfolioPathID = '1tG1ohi0pFMwiZwtSoiAhm';