+++ src/locales/en/portfolioMessages.js
@@ -1,0 +1,7 @@
+// Copyright (c) 2021 Visiosto oy
+// Licensed under the MIT License
+
+export default {
+  portfolioTitle: 'Portfolio:',
+  portfolioCardReadMore: 'Read more',
+};