+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.24.0] - 2021-07-16
+
+### Added
+
+- Text body to the portfolio after the title of the portfolio section on the front page.
+- Portfolio card component to display portfolio entries.
+- Portfolio entries to the front page.
+- Pages for portfolio entries.
+- Property for using lesser paddings in the `Card` components.
+- Support for portfolio pages to the localized links.
+
+### Fixed
+
+- Resizing of the social media icons in the footer on some smaller screens.
+
 ## [0.23.3] - 2021-07-13
 
 ### Fixed
@@ -641,7 +656,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for internationalizing pages.
 - Both light and dark themes for the site.
 
-[unreleased]: https://github.com/visiosto/visiosto-website/compare/v0.23.3...HEAD
+[unreleased]: https://github.com/visiosto/visiosto-website/compare/v0.24.0...HEAD
+[0.24.0]: https://github.com/visiosto/visiosto-website/compare/v0.23.3...v0.24.0
 [0.23.3]: https://github.com/visiosto/visiosto-website/compare/v0.23.2...v0.23.3
 [0.23.2]: https://github.com/visiosto/visiosto-website/compare/v0.23.1...v0.23.2
 [0.23.1]: https://github.com/visiosto/visiosto-website/compare/v0.23.0...v0.23.1